Test of fix for WIN32.
