+++ NEWS.md
@@ -2,7 +2,6 @@
 * Fixed insidious bug that would arise if the edge matrix contained distances > 27.
 * Fixed bug in hdbscan where it would mistakenly conclude that it lacked sufficient neighbors.
 * Change to address apparent upstream issue causing compilation problem on Win32 systems.
-* Bumped version of RcppArmadillo for the same reason.
 
 ### largeVis 0.1.8
 * hdbscan algorithm added
